prefer Chef::ReservedNames::Win32::Version for determining windows version
Per #397 and https://github.com/chef-cookbooks/windows/issues/335, the facilities for determining windows version provided in Chef::ReservedNames::Win32 version are prefered over the helpers provided by the `windows` cookbook's library. My research shows this should work going back to Chef 10, so I am satisfied that this is not likely to be a breaking change.