lazily evaluate sensu_json_file owner and group defaults from node attributes
Prior to this change the default group for a sensu_json_file resource was
hardcoded to 'sensu' as a string.

In [#422](https://github.com/sensu/sensu-chef/pull/422) it is suggested to use the value of node['sensu']['group'] attribute to set this default, but the implementation in that PR is problematic due to the attribute being evaluated when the cookbook loads.

In [#389](https://github.com/sensu/sensu-chef/pull/389) it is suggested to have the sensu_json_file resource manage the user and group for its owner and group, respectively. I'd rather not manage this aspect off the system from what's intended as a simple wrapper around the file resource.

I believe that if we lazily evaluate the 'admin_user' and 'group' attributes as defaults for the sensu_json_file resource we will make it possible for folks to avoid having to create 'sensu' user/group when they are trying to test sensu_json_file resources in isolation.