switch centos-5.10 for centos-5.11
A few things:

The sysv platform suite is failing on both versions of centos 5 because
/usr/sbin is not in root's default $PATH; rabbitmqctl fails to execute
 as a result.

In bdd21cd I added explicit box names for all of our test platforms
because the version of test-kitchen we now need for windows_chef_zero
compatability cannot automatically resolve centos-5.10 to an opscode-
published box name.

Quickly surveying other cookbook projects I see that centos-5.11 is the
version of centos 5 being tested by many. We'll update to that, with the
understanding that this test suite is broken, and exclude it from that suite.
As a result, we can drop the explicit box names.