Support multiple rabbitmq brokers via new `sensu.rabbitmq.hosts` attribute
In general this cookbook passes a great deal of configuration straight through
from node attributes to JSON configuration on disk. Historically this has
included writing rabbitmq transport configuration to disk directly from
attributes as well.

Unfortunately these attributes are somewhat double-bound. They provide both
configuration for Sensu and the parameters for configuring the RabbitMQ
broker on nodes where the `sensu::rabbitmq` recipe is run.

In Sensu 0.15 we added support for multiple rabbitmq brokers in an array,
something which users of this cookbook have heretofore had to manage using
a `sensu_snippet` resource to provide an array of rabbitmq transport
configuration hashes.

With this change we now honor an array of strings as the value of the
 `sensu.rabbitmq.hosts` attribute which, if provided, will be
combined with the other rabbitmq attributes to generate an array of rabbitmq
 transport configuration hashes.

If the `sensu.rabbitmq.hosts` array is empty (as it is by default) we will fall
back to generating an array of rabbitmq transport configuration hashes using
the value provided in `sensu.rabbitmq.host` in combination with any other
rabbitmq attributes.